Default UITableViewController has 0 height table
UITableView.rowHeight is initalized to 0.0.  UITableViewController then uses
this 0.0 rowHeight as a return value for tableView:heightForRowAtIndexPath:

Hence, if you never specify a rowHeight, or if you don't override ...heightForRowAtIndexPath:,
your UITableViewController's table will be 0 pixels tall

Per documentation:
"If you do not explicitly set the row height, UITableView sets it to a standard value."

I'm interpreting this to mean that rowHeight should be initalized to _UITableViewDefaultRowHeight,
rather than 0.0.

UITableViewController implements